projeto_balancetes/requirements.txt
docling
google-genai
pymupdf
python-dotenv
pytest
